+++ inte_prn/admin/__init__.py
@@ -6,6 +6,7 @@ from .end_of_study_admin import EndOfStudyAdmin
 from .integrated_care_clinic_registration_admin import (
     IntegratedCareClinicRegistrationAdmin,
 )
+from .loss_to_followup_admin import LossToFollowupAdmin
 from .onschedule_admin import OnScheduleHivAdmin, OnScheduleNcdAdmin
 from .protocol_deviation_violation_admin import ProtocolDeviationViolationAdmin
 from .unblinding_request_admin import UnblindingRequestAdmin