+++ inte_prn/forms/__init__.py
@@ -2,7 +2,7 @@ from .end_of_study_form import EndOfStudyForm
 from .integrated_care_clinic_registration_form import (
     IntegratedCareClinicRegistrationForm,
 )
-from .loss_to_followup import LossToFollowupForm
+from .loss_to_followup_form import LossToFollowupForm
 from .protocol_deviation_violation_form import ProtocolDeviationViolationForm
 from .unblinding_request_form import UnblindingRequestForm
 from .unblinding_review_form import UnblindingReviewForm