+++ inte_subject/forms/subject_visit_form.py
@@ -1,7 +1,10 @@
+import pdb
+
 from django import forms
-from edc_constants.constants import OTHER, STUDY_DEFINED_TIMEPOINT
+from edc_constants.constants import NOT_APPLICABLE, OTHER, STUDY_DEFINED_TIMEPOINT
 from edc_form_validators import FormValidatorMixin
 from edc_sites.forms import SiteModelFormMixin
+from edc_visit_tracking.constants import MISSED_VISIT, SCHEDULED, UNSCHEDULED
 from edc_visit_tracking.form_validators import VisitFormValidator
 from inte_prn.icc_registered import (
     InterventionSiteNotRegistered,
@@ -26,6 +29,10 @@ class SubjectVisitFormValidator(VisitFormValidator):
         self.validate__clinic_services()
         self.validate__health_services()
 
+        self.applicable_if(
+            SCHEDULED, UNSCHEDULED, field="reason", field_applicable="info_source"
+        )
+
     def validate__clinic_services(self):
         selections = self.get_m2m_selected("clinic_services")
         if (
@@ -42,6 +49,13 @@ class SubjectVisitFormValidator(VisitFormValidator):
             raise forms.ValidationError(
                 {"clinic_services": "This is not a scheduled study visit."}
             )
+
+        self.m2m_applicable_if_true(
+            self.cleaned_data.get("reason") != MISSED_VISIT,
+            m2m_field="clinic_services",
+        )
+
+        self.m2m_single_selection_if(NOT_APPLICABLE, m2m_field="clinic_services")
 
     def validate__health_services(self):
         selections = self.get_m2m_selected("health_services")
@@ -67,6 +81,13 @@ class SubjectVisitFormValidator(VisitFormValidator):
                     }
                 )
         self.m2m_single_selection_if(INTEGRATED, m2m_field="health_services")
+
+        self.m2m_applicable_if_true(
+            self.cleaned_data.get("reason") != MISSED_VISIT,
+            m2m_field="health_services",
+        )
+
+        self.m2m_single_selection_if(NOT_APPLICABLE, m2m_field="health_services")
 
 
 class SubjectVisitForm(SiteModelFormMixin, FormValidatorMixin, forms.ModelForm):